amuleweb: switch session IDs to a 64-bit CSPRNG token (#870)
CScriptWebServer::CheckLoggedin was generating session IDs with C
`rand()` into an `int`, looping on collision. `rand()` is not a CSPRNG,
isn't seeded with anything an attacker can't observe, and yields at
most RAND_MAX (~2^31) candidate values -- making session IDs guessable
in modest time without ever stealing the cookie. The pattern is the
classic "fake session-cookie until you hit a live one and inherit
that user's authenticated context" attack on an admin surface.

Replace with `CryptoPP::AutoSeededRandomPool::GenerateBlock` into a
`uint64_t`:

- 64 bits of entropy is enough that brute-force is no longer a
  realistic attack at amuleweb scale (web request rate, server CPU
  cost per attempt). Going wider (128-bit opaque hex) would also
  work but pulls in cookie-format and storage-type changes for
  marginal improvement.
- AutoSeededRandomPool is the same primitive the EC stack already
  uses for DH key agreement (see EncryptedStreamSocket.cpp), so this
  doesn't drag in a new crypto dependency.
- Reuse the pool across calls via a function-local static so the
  OS-RNG seeding cost is paid once per process, not once per login.

Type ripple (int -> uint64_t):

- `ThreadData::SessionID` (WebServer.h)
- `std::map<int, CSession> m_sessions` -> `std::map<uint64_t, CSession>`
- `CWebSocket::SendHttpHeaders` signature + Set-Cookie format
  (`%d` -> `%llu` with explicit unsigned-long-long cast for the
  windows builds where uint64_t != unsigned long long)
- `int sessid = 0; ... sessid = atoi(...)` -> `uint64_t sessid = 0;
  ... sessid = strtoull(..., 10)` in WebSocket cookie-parse

Closes #870

+++ src/webserver/src/WebServer.h
@@ -724,7 +724,13 @@ class CParsedUrl {
 struct ThreadData {
 	CParsedUrl	parsedURL;
 	wxString	sURL;
-	int		SessionID;
+	// Opaque 64-bit session token; 0 means "no session cookie yet".
+	// Sourced from the project CSPRNG (CryptoPP::AutoSeededRandomPool
+	// via GetRandomPool()) when a new session is created, see
+	// CScriptWebServer::CheckLoggedin in WebServer.cpp. Was `int`
+	// + rand() before #870, which made session IDs trivially
+	// guessable.
+	uint64_t	SessionID;
 	CWebSocket	*pSocket;
 };
 
@@ -834,7 +840,7 @@ class CScriptWebServer : public CWebServerBase {
 		char *GetErrorPage(const char *message, long &size);
 		char *Get_404_Page(long &size);
 
-		std::map<int, CSession> m_sessions;
+		std::map<uint64_t, CSession> m_sessions;
 
 		CSession *CheckLoggedin(ThreadData &);
 	protected:
